[JSX] Imaging Uploader - Refactor and Clean-Up
