+++ modules/imaging_uploader/jsx/ImagingUploader.js
@@ -1,9 +1,9 @@
 import FilterForm from 'FilterForm';
 import {Tabs, TabPane} from 'Tabs';
+import Loader from 'Loader';
 
 import LogPanel from './LogPanel';
 import UploadForm from './UploadForm';
-import formatColumn from './columnFormatter';
 
 class ImagingUploader extends React.Component {
 
@@ -12,13 +12,15 @@ class ImagingUploader extends React.Component {
 
     this.state = {
       isLoaded: false,
-      filter: {}
+      filter: {},
+      hiddenHeaders: ['PatientName']
     };
 
     // Bind component instance to custom methods
-    this.fetchData = this.fetchData.bind(this);
+    this.fetchData    = this.fetchData.bind(this);
     this.updateFilter = this.updateFilter.bind(this);
     this.resetFilters = this.resetFilters.bind(this);
+    this.formatColumn = this.formatColumn.bind(this);
   }
 
   componentDidMount() {
@@ -34,36 +36,159 @@ class ImagingUploader extends React.Component {
     $.ajax(this.props.DataURL, {
       method: "GET",
       dataType: 'json',
-      success: function(data) {
+      success: data => {
         this.setState({
-          Data: data,
+          data: data,
           isLoaded: true
         });
-      }.bind(this),
-      error: function(error) {
-        console.error(error);
-      }
+      },
+      error: error => console.error(error)
     });
   }
 
+  /**
+   * Set this.state.filter to the input filter object
+   *
+   * @param {object} filter = the filter object
+   */
   updateFilter(filter) {
     this.setState({filter});
   }
 
+  //TODO: deprecate clearing filters via refs in future refactoring.
+  /**
+   * Reset the filter elements with textInput refs to empty values
+   */
   resetFilters() {
     this.imagingUploaderFilter.clearFilter();
   }
 
+  /**
+   * Modify behaviour of specified column cells in the Data Table component
+   *
+   * @param {string} column - column name
+   * @param {string} cell - cell content
+   * @param {arrray} rowData - array of cell contents for a specific row
+   * @param {arrray} rowHeaders - array of table headers (column names)
+   *
+   * @return {*} a formated table cell for a given column
+   */
+  formatColumn(column, cell, rowData, rowHeaders) {
+    // If a column if set as hidden, don't display it
+    if (this.state.hiddenHeaders.indexOf(column) > -1) {
+      return null;
+    }
+  
+    // Create the mapping between rowHeaders and rowData in a row object.
+    const row = {};
+    rowHeaders.forEach((header, index) => row[header] = rowData[index]);
+  
+    // Default cell style
+    const cellStyle = {whiteSpace: 'nowrap'};
+  
+    if (column === 'Progress') {
+      if (cell === 'Failure') {
+        cellStyle.color = '#fff';
+        return (
+          <td className="label-danger" style={cellStyle}>
+            {cell}
+          </td>
+        );
+      }
+  
+      if (cell === 'In Progress...') {
+        cellStyle.color = '#fff';
+        return (
+          <td className="label-warning" style={cellStyle}>
+            {cell}
+          </td>
+        );
+      }
+  
+      const created = row['Number Of MincCreated'];
+      const inserted = row['Number Of MincInserted'];
+      return (
+        <td style={cellStyle}>
+          {cell} ({inserted} out of {created})
+        </td>
+      );
+    }
+  
+    if (column === 'Tarchive Info') {
+      if (!cell || cell === "0") {
+        return (<td></td>);
+      }
+  
+      const url = loris.BaseURL + '/dicom_archive/viewDetails/?tarchiveID=' + cell;
+      return (
+        <td style={cellStyle}>
+          <a href={url}>View Details</a>
+        </td>
+      );
+    }
+  
+    if (column === 'Number Of MincInserted') {
+      if (cell > 0) {
+        return (
+          <td style={cellStyle}>
+            <a onClick={this.handleClick.bind(null, row.CandID)}>{cell}</a>
+          </td>
+        );
+      }
+    }
+  
+    if (column === 'Number Of MincCreated') {
+      let violatedScans;
+      if (row['Number Of MincCreated'] - row['Number Of MincInserted'] > 0) {
+        let numViolatedScans =
+             row['Number Of MincCreated'] - row['Number Of MincInserted'];
+  
+        let patientName = row.PatientName;
+        violatedScans = <a onClick={this.openViolatedScans.bind(null, patientName)}>
+           ({numViolatedScans} violated scans)
+         </a>;
+      }
+  
+      return (
+        <td style={cellStyle}>
+          {cell}
+          &nbsp;
+          {violatedScans}
+        </td>
+      );
+    }
+
+    return (<td style={cellStyle}>{cell}</td>);
+  }
+  
+  /**
+   * Handles clicks on 'Number Of MincInserted' cells
+   *
+   * @param {string} dccid - dccid
+   * @param {object} e - event info
+   */
+  handleClick(dccid, e) {
+    loris.loadFilteredMenuClickHandler('imaging_browser/', {
+      DCCID: dccid
+    })(e);
+  }
+  
+  /**
+   * Opens MRI Violations for when there are violated scans
+   *
+   * @param {string} patientName - Patient name of the form PSCID_DCCID_VisitLabel
+   * @param {object} e - event info
+   */
+  openViolatedScans(patientName, e) {
+    loris.loadFilteredMenuClickHandler('mri_violations/', {
+      PatientName: patientName
+    })(e);
+  }
+  
+
   render() {
     if (!this.state.isLoaded) {
-      return (
-        <button className="btn-info has-spinner">
-          Loading
-          <span
-            className="glyphicon glyphicon-refresh glyphicon-refresh-animate">
-          </span>
-        </button>
-      );
+      return <Loader/>;
     }
 
     const tabList = [
@@ -71,6 +196,7 @@ class ImagingUploader extends React.Component {
       {id: "upload", label: "Upload"}
     ];
 
+    //TODO: deprecate use of filter refs in future refactor.
     const filterRef = function(f) {
       this.imagingUploaderFilter = f;
     }.bind(this);
@@ -88,9 +214,9 @@ class ImagingUploader extends React.Component {
                 onUpdate={this.updateFilter}
                 filter={this.state.filter}
               >
-                <TextboxElement {... this.state.Data.form.candID} />
-                <TextboxElement {... this.state.Data.form.pSCID} />
-                <SelectElement {... this.state.Data.form.visitLabel} />
+                <TextboxElement {... this.state.data.form.candID} />
+                <TextboxElement {... this.state.data.form.pSCID} />
+                <SelectElement {... this.state.data.form.visitLabel} />
                 <ButtonElement type="reset" label="Clear Filters" onUserInput={this.resetFilters}/>
               </FilterForm>
             </div>
@@ -100,18 +226,19 @@ class ImagingUploader extends React.Component {
           </div>
           <div id="mri_upload_table">
             <StaticDataTable
-              Data={this.state.Data.Data}
-              Headers={this.state.Data.Headers}
-              getFormattedCell={formatColumn}
+              Data={this.state.data.Data}
+              Headers={this.state.data.Headers}
+              hiddenHeader={this.state.hiddenHeaders}
+              getFormattedCell={this.formatColumn}
               Filter={this.state.filter}
             />
           </div>
         </TabPane>
         <TabPane TabId={tabList[1].id}>
           <UploadForm
-            form={this.state.Data.form}
-            mriList={this.state.Data.mriList}
-            maxUploadSize={this.state.Data.maxUploadSize}
+            form={this.state.data.form}
+            mriList={this.state.data.mriList}
+            maxUploadSize={this.state.data.maxUploadSize}
           />
         </TabPane>
       </Tabs>
@@ -119,7 +246,8 @@ class ImagingUploader extends React.Component {
   }
 }
 
-ImagingUploader.propTypes = {};
-ImagingUploader.defaultProps = {};
+ImagingUploader.propTypes = {
+  DataURL: React.PropTypes.string.isRequired
+};
 
 export default ImagingUploader;
